Unsubscribe from newsletter :sparkles:

+++ api/newsletter/delete.php
@@ -21,16 +21,30 @@ $db = $database->connect();
 $newsletter = new Newsletter($db);
 
 //Assign newsletter properties if param is valid
-$newsletter->id = validate_param($_GET['id']);
+$newsletter->id = isset($_GET['id']) ? strip_tags($_GET['id']) : 'null';
+$newsletter->email = isset($_GET['email']) ? strip_tags($_GET['email']) : 'null';
 
-//Try to delete the newsletter
-if($newsletter->delete()){
-  echo json_encode(array(
-    'success' => 'L\'adresse email a bien été supprimé de notre newsletter !'
-  ));
+$id_count = $db->query('SELECT * FROM newsletter WHERE id=' . $newsletter->id)->rowCount();
+$email_count = $db->query('SELECT * FROM newsletter WHERE email="' . $newsletter->email  . '"')->rowCount();
+
+if($id_count==1 || $email_count==1){
+  //Check that the id or the email isn't empty
+  if(!empty($newsletter->id) || !empty($newsletter->email)){
+    //Try to delete the newsletter
+    if($newsletter->delete()){
+      echo json_encode(array(
+        'success' => 'L\'adresse email a bien été supprimé de notre newsletter !'
+      ));
+    }else{
+      HTTPStatus(500);
+      echo json_encode(array(
+        'error' => 'Une erreur est survenue'
+      ));
+    }
+  }
 }else{
-  HTTPStatus(500);
+  HTTPStatus(404);
   echo json_encode(array(
-    'error' => 'Une erreur est survenue'
+    'error' => 'L\'adresse email n\'est pas inscrite à notre newsletter'
   ));
 }
